Move ServletHelper into util package.
